Optimize Mixer
## Modified synthesis.c
- Fixed ordering on DMEM_ADDR defines.
- Improved documentation and naming.
- Split synthesis_process_note(s) into multiple versions: one for N64, and two for non-N64. The original version was extremely difficult to read and work with due to the maze of preprocessor directives.
- Added support for Enhanced RSPA Emulation (see mixer.c heading for more info).
- Documented most of synthesis_process_note(s). The US-non-N64 version has been rearranged for best readability.
- Added some profiling info (will clean up before merge).

## Modified level_script.c
- Added commented-out 3DS sleeps.

## Modified mixer.c (RSP Audio Emulation)
- Removed all functional code and moved it to different files inside of the mixer_implementations directory. This file now selects one of those other files at build-time.

## Modified mixer.h
- Added Enhanced RSPA Emulation build-time toggle to bypass some inefficiencies that were necessary on N64, but unnecessary in emulation.
- Improved documentation.

## Created mixer_neon.c, mixer_sse41.c, mixer_reference.c, mixer_3ds.c
- These files contain complete RSPA implementations.

## Created mixer_3ds.c
- Contains an RSPA implementation optimized for 3DS.
- Added additional clamping functions. On 3DS, it's significantly faster to break clamping into two steps, which causes GCC to use conditional math instead of conditional branches.
- Rewrote aEnvMixerImpl and aMixImpl for better 3DS performance and readability.

## Created profiler_3ds.c, profiler_3ds.h
- Used as a quick-and-dirty 3DS-specific profiling system intended for use with a debugger.
- Can be enabled or disabled at build time with a one-line edit to the header file.
- History sizes can be configured.
- Notably, does not use any 3DS includes, and thus does not destroy your type definitions.

## Modified audio_3ds_threading.h
- Added support for the 3DS threadsleep function, which can be enabled/disabled at compile time.
- Renamed some stuff to be more consistent.

## Modified audio_3ds.c
- Now uses the 3DS sleep function in audio_3ds_threading.h when appropriate.
- Now uses the profiler.

### Files with updated documentation only
- controller_3ds.c
- gfx_3ds.h
