+++ src/pc/mixer_implementations/note.txt
@@ -1,0 +1,1 @@
+This directory must not be included in the build path. The correct file is selected dynamically by mixer.c.
